Add MySQL-python dependency
This is to allow other packages that already depend on python-oslo-db to
get rid of that dependency.

Note that technically speaking oslo.db does not require the package too,
because we have plenty of options to deploy OpenStack, using other
database backends different from MySQL, or just using other Python
drivers for SQLAlchemy.

Ideally, this added dependency should be dropped after a while, when all
installers install the proper Python driver for the type of setup
they deploy.

Change-Id: I1d1852c993e6e1a28fbe9e0f693864471213c18e