Update patch set 2
Patch Set 2:

This is intended to fix https://bugzilla.redhat.com/show_bug.cgi?id=1566280 . The real problem is that OSP is not longer shipping oslo-db-tests subpackage as it's not longer needed for tempest and that's breaking upgrade path. This is a workaround, we could figure out something better.

Patch-set: 2