Remove MySQL-python as a requirement
Long time ago, the OpenStack community switched to PyMySQL as the
standard library to use for MySQL access [1]. Before that, we had
MySQL-python, and it was kept as a runtime requirement for the
oslo-db package for backwards compatibility reasons.

Since a long time has passed, let's remove the MySQL-python requirement
completely, we have been using PyMySQL for our deployments for more
than two years.

[1] - https://trello.com/c/q0VoAYJq/89-migrate-mysql-python-to-pymysql?menu=filter&filter=label:RDO-Infra

Change-Id: I25dd065f00c061fefc3fc7abab4e84298bfd0d7d