Add python3-pynacl dep to support auth_ed25519 in pymysql
oslo.db clients connect to the mysql database with pymysql.
When a stack is configured to use auth_ed25519 for client
authentication, pymysql requires package python3-pynacl for
the crypto bits needed to implement the authentication.

Add a new dep on python3-pynacl to make the new authentication
method always available to oslo.db clients.

Tested by deploying a standalone cloud with TripleO.

Change-Id: Id5fc620f2ed540e15233c1180b72cc4d57f88f5a
Closes-Bug: #1898263